requirements.txt
# Core LangChain and LangGraph
langchain>=0.2.0
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# Web Search
tavily-python>=0.3.0

# PDF and Document Processing
pypdf>=4.0.0
pymupdf>=1.24.0
arxiv>=2.1.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
httpx>=0.27.0

# Optional: For better async support
aiohttp>=3.9.0
